Configure cloudadmin as the admin user for EC2 instances
This project is already using the 'cloudadmin' user as the admin user
for remote systems in Azure and GCP, so AWS/EC2 was the outlier by
having the 'ec2-user' provided by default in EC2 as the admin user.
This commit adds 'cloudadmin' as the admin user instead of
'ec2-user' via the user_data option in the AWS EC2 instances
configuration. At the same time, 'ec2-user' is being dropped from the
remote systems as a consequence.

user_data configuration for 'cloudadmin' as a substitute to 'ec2-user'
was taken from: https://aws.amazon.com/premiumsupport/knowledge-center/ec2-user-account-cloud-init-user-data/