Allow qesap.py to parse lines with more than one space in tfvars templates (#138)
Current version of the glue script expect 0 or 1 spaces between a
variable and the equal sign that follows while parsing the lines from
the tfvars template file to properly identify the variable name and its
value. This causes problems when more than one space is used, as in L5
from https://github.com/SUSE/qe-sap-deployment/blob/main/terraform/azure/terraform.tfvars.template
which is:

vnet_address_range   = "10.10.0.0/16"

Lines such as this could cause variables defined in such a way to be
duplicated in the resulting terraform.tfvars file, causing terraform to
fail due to duplicate variable definitions.

This commit introduces a fix to allow 0 or mare spaces between the
variable name and the equal sign.