Add workaround for bsc1224249
S:M:33974:python3 breaks installation of `python3` in 15-SP5 images in
CSP's catalogs. Seems issue comes from the `python3-argcomplete` package
shipped in the OS image, which breaks a post installation script during
the installation of `python3`. This workaround will only be required if
S:M:33974 gets released, and until 15-SP5 images are refreshed in the
CSPs later in 2024.