Add cloud provider name in the intentory
On the Terraform side: add the name of the cloud provider to the inventory.
On the Ansible side: make the platform detection optional only in case
the variable is not configured from the external (aka the inventory)
This change allow to save time in many ansible playbooks.
fix some Ansible lint warnings.