Install fence-agents-azure-arm explicitly.
Workaround for bsc#1224797: bug was opened because the update of
`python3-azure-sdk` to `python311-azure-sdk` caused `fence_azure_arm` to
fail to start due to missing dependencies. New package
`fence-agents-azure-arm` was created to supply to impacted service packs
a native Azure fence agent compatible with `python311-azure-sdk`, but
this package needs to be manually installed as it was not shipped as
part of the base OS image.