Change Terraform GCP module API aboput machinetype
Change variable name in all module from machine_type to vm_size.
Adapt accordingly the module main and the top level main.
Add configurability for Monitor, the only one was missing a variable for
the machine type.
Remove all the default values for vm_size, at module level.
This commit does not result in any API or functional change for the end
user.