Make AWS/GCP HANA cluster playbook follow best practices
In some scenarios, playbooks to configure HANA resources in a cluster
are failing to deploy a healthy HanaSR cluster in AWS. This could be
either due to a race condition, or to not configuring the clone resource
into the cluster in maintenance mode as specified in the best practices.
This commits adds into the `cluster-hana.yaml` task calls to
`cs_wait_for_idle` after commands which modify the cluster
configuration, as well as modifying the cluster configuration so the
SAPHana resource clone is added in maintenance mode. Also calls were
added to remove the maintenance flag from this resource later in the
playbook.