Change retry timing for SUSEConnect
Retry less time but wait double the time between each retry. Without
this change, the retry is orking properly in detecting the error and run
the command again, in particular zypper error 7.
Problem is, in PAYG images, the first retry that does not fails for a
collision with another zypper process, report some modules as not
Registered: it is something not expected for a PAYG image.
Waiting more time can give time to the image itself (cloudregister) to
register all modules.