Tune ansible lint configuration
Let's run the github workflow to use a version of ansible-lint that is
compatible to the official ansible-core version used by the project.
Adjust yamllint configuration to what used by ansible.
Lower fatal ansible-lint rules about roles path: it is something about
the qe-sap-deployment structure that needs more works to be addressed.