e2e test restructuring
Split the script in smaller scripts, each one for a test group,
one for each glue script sub-command:
- configure
- terraform
- ansible
- deployment

The test.sh remain the same from the user perspective, so no changes
are needed in the documentation or any github workflow.
But now it is also possible to manually run it like:
  `./test.sh ansible`
and only have subset of tests executed.

This commit also contain:
- fix in test name printing
- new test about variables in the Ansible section
- fix for sudo: a recent commit in the gluescript
(ref. 55d3714036db874fe33a2f381dd5983866707d6a )
introduced the execution of `sudo -n true` command.
This is problematic for systems that has password protected sudo.
This is not usually the case when running the deployment
with sles4sap cloud images, but become problematic when running
the e2e test locally.